pyproject.toml
[tool.poetry]
name = "gate"
version = "0.1.0"
description = ""
authors = ["dyuzha <matvey10d@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.28.1"
amocrm-api = "^2.6.1"
uvicorn = "^0.38.0"
pydantic-settings = "^2.12.0"
fastapi = "^0.123.9"
pytest-dotenv = "^0.5.2"

[tool.poetry.group.test.dependencies]
pytest = "^9.0.1"
pytest-asyncio = "^1.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-s --log-cli-level=INFO"
env_files = ".test.env"
testpaths = ["tests"]
python_files = ["test_*.py"]
filterwarnings = [
    "ignore::DeprecationWarning"
]
markers = [
    "integration: интеграционные тесты"
]
